remove last remanant of `RawImage`
