+++ crates/egui/src/load/texture_loader.rs
@@ -1,0 +1,60 @@
+use super::*;
+
+#[derive(Default)]
+pub struct DefaultTextureLoader {
+    cache: Mutex<HashMap<(String, TextureOptions), TextureHandle>>,
+}
+
+impl TextureLoader for DefaultTextureLoader {
+    fn id(&self) -> &str {
+        crate::generate_loader_id!(DefaultTextureLoader)
+    }
+
+    fn load(
+        &self,
+        ctx: &Context,
+        uri: &str,
+        texture_options: TextureOptions,
+        size_hint: SizeHint,
+    ) -> TextureLoadResult {
+        let mut cache = self.cache.lock();
+        if let Some(handle) = cache.get(&(uri.into(), texture_options)) {
+            let texture = SizedTexture::from_handle(handle);
+            Ok(TexturePoll::Ready { texture })
+        } else {
+            match ctx.try_load_image(uri, size_hint)? {
+                ImagePoll::Pending { size } => Ok(TexturePoll::Pending { size }),
+                ImagePoll::Ready { image } => {
+                    let handle = ctx.load_texture(uri, image, texture_options);
+                    let texture = SizedTexture::from_handle(&handle);
+                    cache.insert((uri.into(), texture_options), handle);
+                    Ok(TexturePoll::Ready { texture })
+                }
+            }
+        }
+    }
+
+    fn forget(&self, uri: &str) {
+        #[cfg(feature = "log")]
+        log::trace!("forget {uri:?}");
+
+        self.cache.lock().retain(|(u, _), _| u != uri);
+    }
+
+    fn forget_all(&self) {
+        #[cfg(feature = "log")]
+        log::trace!("forget all");
+
+        self.cache.lock().clear();
+    }
+
+    fn end_frame(&self, _: usize) {}
+
+    fn byte_size(&self) -> usize {
+        self.cache
+            .lock()
+            .values()
+            .map(|texture| texture.byte_size())
+            .sum()
+    }
+}
